requirements.txt
Django==1.10.1
django-taggit==0.21.2
djangorestframework==3.4.7
